+++ crates/egui-wgpu/src/winit.rs
@@ -117,7 +117,9 @@ impl<'a> Painter<'a> {
 
         if self.render_state.is_none() {
             let adapter = self.adapter.as_ref().unwrap();
-            let swapchain_format = surface.get_supported_formats(adapter)[0];
+
+            let swapchain_format =
+                select_framebuffer_format(&surface.get_supported_formats(adapter));
 
             let rs = pollster::block_on(self.init_render_state(adapter, swapchain_format));
             self.render_state = Some(rs);
@@ -321,4 +323,16 @@ impl<'a> Painter<'a> {
     pub fn destroy(&mut self) {
         // TODO(emilk): something here?
     }
+}
+
+fn select_framebuffer_format(formats: &[wgpu::TextureFormat]) -> wgpu::TextureFormat {
+    for &format in formats {
+        if matches!(
+            format,
+            wgpu::TextureFormat::Rgba8Unorm | wgpu::TextureFormat::Bgra8Unorm
+        ) {
+            return format;
+        }
+    }
+    formats[0] // take the first
 }