+++ crates/egui_glium/examples/native_texture.rs
@@ -115,7 +115,7 @@ fn create_display(event_loop: &glutin::event_loop::EventLoop<()>) -> glium::Disp
 
     let context_builder = glutin::ContextBuilder::new()
         .with_depth_buffer(0)
-        .with_srgb(true)
+        .with_srgb(false)
         .with_stencil_buffer(0)
         .with_vsync(true);