+++ CHANGELOG.md
@@ -6,6 +6,10 @@ NOTE: [`epaint`](crates/epaint/CHANGELOG.md), [`eframe`](crates/eframe/CHANGELOG
 
 ## Unreleased
 * ⚠️ BREAKING: Fix text being too small ([#2069](https://github.com/emilk/egui/pull/2069)).
+* ⚠️ BREAKING: egui now expects integrations to do all color blending in gamma space ([#2071](https://github.com/emilk/egui/pull/2071)).
+
+### Fixed 🐛
+* Improved text rendering ([#2071](https://github.com/emilk/egui/pull/2071)).
 
 
 ## 0.19.0 - 2022-08-20