+++ crates/epaint/CHANGELOG.md
@@ -4,6 +4,7 @@ All notable changes to the epaint crate will be documented in this file.
 
 ## Unreleased
 * ⚠️ BREAKING: Fix text being too small ([#2069](https://github.com/emilk/egui/pull/2069)).
+* ⚠️ BREAKING: epaint now expects integrations to do all color blending in gamma space ([#2071](https://github.com/emilk/egui/pull/2071)).
 
 
 ## 0.19.0 - 2022-08-20